+++ web/html/src/manager/virtualization/nets/edit/nets-edit.tsx
@@ -1,0 +1,54 @@
+import { ActionChain } from "components/action-schedule";
+
+import * as React from "react";
+import { TopPanel } from "components/panels/TopPanel";
+import { Loading } from "components/utils/Loading";
+import { SimpleActionApi } from "../../SimpleActionApi";
+import { VirtualizationNetworkDefinitionApi } from "../virtualization-network-definition-api";
+import { NetworkProperties } from "../network-properties";
+
+type Props = {
+  serverId: string;
+  netName: string;
+  localTime: string;
+  timezone: string;
+  actionChains: Array<ActionChain>;
+};
+
+export function NetsEdit(props: Props) {
+  return (
+    <SimpleActionApi
+      urlType="nets"
+      idName="names"
+      hostid={props.serverId}
+      bounce={`/rhn/manager/systems/details/virtualization/nets/${props.serverId}`}
+    >
+      {({ onAction, messages: actionMessages }) => (
+        <VirtualizationNetworkDefinitionApi hostid={props.serverId} networkName={props.netName}>
+          {({ definition, messages: definitionMessages }) => {
+            if (definition == null) {
+              return <Loading />;
+            }
+
+            const onSubmit = (model: any) => onAction("edit", [model.definition.name], model);
+
+            return (
+              <TopPanel title={props.netName}>
+                <NetworkProperties
+                  serverId={props.serverId}
+                  submitText={t("Update")}
+                  submit={onSubmit}
+                  initialModel={definition}
+                  messages={actionMessages.concat(definitionMessages)}
+                  timezone={props.timezone}
+                  localTime={props.localTime}
+                  actionChains={props.actionChains}
+                />
+              </TopPanel>
+            );
+          }}
+        </VirtualizationNetworkDefinitionApi>
+      )}
+    </SimpleActionApi>
+  );
+}
